refactor: OrderStateRestorable
